WIP F1 25: 2026 season pack UDP

+++ src/PacketParticipantData.h
@@ -16,9 +16,9 @@ struct LiveryColour {
 struct ParticipantData
 {
     uint8_t m_aiControlled; // Whether the vehicle is AI (1) or Human (0) controlled
-    uint8_t m_driverId; // Driver id - see appendix, 255 if network human
-    uint8_t m_networkId; // Network id – unique identifier for network players
-    uint8_t m_teamId; // Team id - see appendix
+    uint16_t m_driverId; // Driver id - see appendix, 65535 if network human
+    uint16_t m_networkId; // Network id – unique identifier for network players
+    uint16_t m_teamId; // Team id - see appendix
     uint8_t m_myTeam; // My team flag – 1 = My Team, 0 = otherwise
     uint8_t m_raceNumber; // Race number of the car
     uint8_t m_nationality; // Nationality of the driver
@@ -51,4 +51,3 @@ private:
 #pragma pack(pop)
 
 #endif 
-
